feature: post #10 마커 불러오기 될 때 마커에 해당하는 포스트들 정보 불러오기 성공

+++ app/src/main/java/com/example/trace_android/model/Post.kt
@@ -16,7 +16,7 @@ data class PostResponse(
 
 // 서버에서 반환되는 포스트 데이터의 구조를 나타내는 클래스
 data class Post(
-    val id: Int,                        //포스팅의 id
+    val id: Long,                        //포스팅의 id
     val email: String,                  //작성자의 Email
     val content: String,
     val latitude: Double,
